requirements.txt
python-dotenv==0.20.0
requests==2.28.1
pandas==1.4.3
dash==2.5.1
plotly==5.9.0
gunicorn==20.1.0
uvicorn==0.18.2
seaborn==0.11.2
Counter==1.0.0